+++ lib/smart_vk_api/configuration.rb
@@ -1,0 +1,7 @@
+module SmartVkApi
+  class Configuration
+    attr_accessor :app_id
+    attr_accessor :app_token
+    attr_accessor :access_token
+  end
+end