+++ spec/smart_vk_api_spec.rb
@@ -1,0 +1,24 @@
+require 'spec_helper'
+
+describe SmartVkApi do
+  describe 'configuration' do
+    it { is_expected.to respond_to(:configure) }
+
+    before do
+      SmartVkApi.configure do |conf|
+        conf.app_id = '0000000'
+        conf.app_token = '7a6fa4dff77a228eeda56603'
+      end
+    end
+
+    it 'should allow to reset configuration' do
+      SmartVkApi.reset_configuration
+      expect(SmartVkApi.configuration).to be nil
+    end
+
+    it 'should allow to read global configuration' do
+      expect(SmartVkApi.configuration.app_id).to eq('0000000')
+      expect(SmartVkApi.configuration.app_token).to eq('7a6fa4dff77a228eeda56603')
+    end
+  end
+end